Fix login and signup page
