+++ server.js
@@ -6,36 +6,12 @@ const session = require('express-session');
 const exphbs = require('express-handlebars');
 const SequelizeStore = require('connect-session-sequelize')(session.Store);
 
-
 const routes = require('./controllers');
 const sequelize = require('./config/connection');
 const helpers = require('./utils/helpers');
 
 const app = express();
 const PORT = process.env.PORT || 3001;
-
-// Set up sessions
-const sess = {
-  secret: 'Super secret secret',
-  cookie: {
-    // Stored in milliseconds
-    maxAge: 24 * 60 * 60 * 1000, // expires after 1 day
-
-const path = require('path');
-const express = require('express');
-const session = require('express-session');
-const exphbs = require('express-handlebars');
-const routes = require('./controllers');
-const helpers = require('./utils/helpers');
-
-const sequelize = require('./config/connection');
-const SequelizeStore = require('connect-session-sequelize')(session.Store);
-
-const app = express();
-const PORT = process.env.PORT || 3001;
-
-// Set up Handlebars.js engine with custom helpers
-const hbs = exphbs.create({ helpers });
 
 const sess = {
   secret: 'Super secret secret',
@@ -50,13 +26,11 @@ const sess = {
   store: new SequelizeStore({
     db: sequelize,
   }),
-    db: sequelize
-  })
 };
 
 app.use(session(sess));
 
-
+// Set up Handlebars.js engine with custom helpers
 const hbs = exphbs.create({ helpers });
 
 // Inform Express.js on which template engine to use
@@ -71,14 +45,9 @@ app.use(express.static(path.join(__dirname, 'public')));
 app.use(routes);
 
 sequelize.sync({ force: false }).then(() => {
-
-  app.listen(PORT, () => 
-  console.log(
-    `\nServer running on port ${PORT}. Visit http://localhost:${PORT} and create an account!`
+  app.listen(PORT, () =>
+    console.log(
+      `\nServer running on port ${PORT}. Visit http://localhost:${PORT} and create an account!`
     )
   );
 });
-
-  app.listen(PORT, () => console.log('Now listening'));
-});
-
