+++ views/login.handlebars
@@ -1,8 +1,8 @@
-<div class="flex h-full">
+<div class="flex h-full p-6">
   <div class="w-1/2 p-12">
     <h2 class="text-xl mb-6 font-bold">Log in</h2>
 
-    <form class="w-full">
+    <form class="w-full" id="login-form">
       <div class="mb-3">
         <label class="block p-0 font-semibold text-sm mb-1" for="email-login">Email</label>
         <input class="text-sm p-2 pl-3 rounded-lg w-full border border-gray-300 focus:outline-emerald-600" type="text" id="email-login" placeholder="example@gmail.com" />
@@ -12,34 +12,14 @@
         <input class="text-sm p-2 pl-3 rounded-lg w-full border border-gray-300 focus:outline-emerald-600" type="password" id="password-login" placeholder="Password" />
       </div>
       <div class="mt-6">
-        <button class="bg-emerald-200 font-semibold w-full p-2 pl-6 pr-6 rounded-lg shadow-lg" type="submit">Login</button>
+        <button class="bg-emerald-700 text-neutral-50 font-semibold w-full p-2 pl-6 pr-6 rounded-lg shadow-lg" type="submit">Login</button>
       </div>
     </form>
-  </div>
-  <div class="w-1/2 bg-[url('/images/login-splash.jpg')] bg-cover bg-no-repeat">
-    {{!-- <img src="/images/login-splash.jpg" /> --}}
-  </div>
-  {{!-- <div class="col-md-6">
-    <h2>Signup</h2>
 
-    <form class="form signup-form">
-      <div class="form-group">
-        <label for="name-signup">name:</label>
-        <input class="form-input" type="text" id="name-signup" />
-      </div>
-      <div class="form-group">
-        <label for="email-signup">email:</label>
-        <input class="form-input" type="text" id="email-signup" />
-      </div>
-      <div class="form-group">
-        <label for="password-signup">password:</label>
-        <input class="form-input" type="password" id="password-signup" />
-      </div>
-      <div class="form-group">
-        <button class="btn btn-primary" type="submit">signup</button>
-      </div>
-    </form>
-  </div> --}}
+    <a href="/signup" class="block text-emerald-700 mt-3">Don't have an account? Sign up for free.</a>
+  </div>
+  <div class="w-1/2 bg-[url('/images/login-splash.jpg')] bg-cover bg-no-repeat rounded-tl-3xl rounded-br-3xl">
+  </div>
 </div>
 
 <script src="./js/login.js"></script>
