+++ views/signup.handlebars
@@ -1,0 +1,29 @@
+<div class="flex h-full p-6">
+  <div class="w-1/2 p-12">
+    <h2 class="text-xl mb-6 font-bold">Signup</h2>
+
+    <form class="w-full" id="signup-form">
+			<div class="mb-3">
+        <label class="block p-0 font-semibold text-sm mb-1" for="name-signup">Name</label>
+        <input class="text-sm p-2 pl-3 rounded-lg w-full border border-gray-300 focus:outline-emerald-600" type="text" id="name-signup" placeholder="Natasa K" />
+      </div>
+			<div class="mb-3">
+        <label class="block p-0 font-semibold text-sm mb-1" for="email-signup">Email</label>
+        <input class="text-sm p-2 pl-3 rounded-lg w-full border border-gray-300 focus:outline-emerald-600" type="text" id="email-signup" placeholder="example@gmail.com" />
+      </div>
+			<div class="mb-3">
+        <label class="block p-0 font-semibold text-sm mb-1" for="password-signup">Password</label>
+        <input class="text-sm p-2 pl-3 rounded-lg w-full border border-gray-300 focus:outline-emerald-600" type="password" id="password-signup" />
+      </div>
+      <div class="mt-6">
+        <button class="bg-emerald-700 text-neutral-50 font-semibold w-full p-2 pl-6 pr-6 rounded-lg shadow-lg" type="submit">Sign up</button>
+      </div>
+    </form>
+
+		<a href="/login" class="block text-emerald-700 mt-3">Already have an account? Log in here.</a>
+  </div>
+  <div class="w-1/2 bg-[url('/images/login-splash.jpg')] bg-cover bg-no-repeat rounded-tl-3xl rounded-br-3xl">
+  </div>
+</div>
+
+<script src="./js/signup.js"></script>
